feat: use JWT/JWKS to authorize blob operations
Initial draft for requiring JSON Web Tokens to access the blobs, validating
against a JWKS retrieved from an issuer. At this time, the "jwks_uri" is
assumed to be the ISSUER_URI + "/.well-known/jwks.json" but that is not
standard, and hence will be changed in the near future.

cargo test passes